content/html/content/test/file_fullscreen-svg-element.html

<!DOCTYPE HTML>
<html>
  <!--
  https://bugzilla.mozilla.org/show_bug.cgi?id=735031
  Bug 735031 - Fullscreen API implementation assumes an HTML Element
  -->
  <head>
    <title>Bug 735031</title>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  </head>
  <body>
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=73503">
      Mozilla Bug 735031</a>

    <svg width="100" height="100" viewbox="0 0 100 100">
        <rect id="svg-elem" x="10" y="10" width="50" height="50"
        fill="black" stroke="blue" stroke-width="2"/>
    </svg>

    <pre id="test">
      <script type="application/javascript">
        /*
         * Test for Bug 735031
         * Test locking non-html element.
         */
        function begin() {
          var elem = document.getElementById("svg-elem")
            , elemWasLocked = false;

          document.addEventListener("mozfullscreenchange", function (e) {
            if (document.mozFullScreenElement === elem) {
              elemWasLocked = true;
              document.mozCancelFullScreen();
            } else {
              opener.ok(elemWasLocked, "Expected SVG elem to become locked.");
              opener.nextTest();
            }
          }, false);
          elem.mozRequestFullScreen();
        }
      </script>
    </pre>
  </body>
</html>
